# internal/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"AI News Mailer","description":"A REST service to interact with an emailer that sends AI-summarised news","version":"1.0"},"host":"localhost:8888","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/subscription":{"get":{"tags":["subscription"],"summary":"getAll subscription","description":"Get all subscriptions","operationId":"subscription#getAll","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/SubscriptionResponse"}}}},"schemes":["http"]},"post":{"tags":["subscription"],"summary":"createOne subscription","description":"Create a new subscription","operationId":"subscription#createOne","parameters":[{"name":"CreateOneRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/SubscriptionCreateOneRequestBody","required":["email"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/SubscriptionCreateOneResponseBody"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/SubscriptionCreateOneSubscriptionFieldMissingResponseBody"}},"409":{"description":"Conflict response.","schema":{"$ref":"#/definitions/SubscriptionCreateOneSubscriptionAlreadyExistsResponseBody"}}},"schemes":["http"]}},"/subscription/{uid}":{"get":{"tags":["subscription"],"summary":"getOneByUID subscription","description":"Get a subscription by UID","operationId":"subscription#getOneByUID","parameters":[{"name":"uid","in":"path","description":"UID of subscription","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/SubscriptionGetOneByUIDResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/SubscriptionGetOneByUIDSubscriptionNotFoundResponseBody"}}},"schemes":["http"]},"delete":{"tags":["subscription"],"summary":"deleteOneByUID subscription","description":"Delete a subscription by UID","operationId":"subscription#deleteOneByUID","parameters":[{"name":"uid","in":"path","description":"UID of subscription","required":true,"type":"string"}],"responses":{"200":{"description":"OK response."},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/SubscriptionDeleteOneByUIDSubscriptionNotFoundResponseBody"}}},"schemes":["http"]}}},"definitions":{"SubscriptionCreateOneRequestBody":{"title":"SubscriptionCreateOneRequestBody","type":"object","properties":{"email":{"type":"string","description":"Subscribed email","example":"user@email.com"},"name":{"type":"string","description":"Name of the subscriber","example":"John Doe"},"since":{"type":"string","description":"Datetime when the subscription was made","example":"2012-04-23T18:25:43.511Z"},"uid":{"type":"string","description":"Unique identifier of subcription","example":"abcd1234"}},"description":"Subscription to be created","example":{"email":"user@email.com","name":"John Doe","since":"2012-04-23T18:25:43.511Z","uid":"abcd1234"},"required":["email"]},"SubscriptionCreateOneResponseBody":{"title":"SubscriptionCreateOneResponseBody","type":"object","properties":{"email":{"type":"string","description":"Subscribed email","example":"user@email.com"},"name":{"type":"string","description":"Name of the subscriber","example":"John Doe"},"since":{"type":"string","description":"Datetime when the subscription was made","example":"2012-04-23T18:25:43.511Z"},"uid":{"type":"string","description":"Unique identifier of subcription","example":"abcd1234"}},"description":"Created subscription","example":{"email":"user@email.com","name":"John Doe","since":"2012-04-23T18:25:43.511Z","uid":"abcd1234"}},"SubscriptionCreateOneSubscriptionAlreadyExistsResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Error returned when the subscription already exists (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"SubscriptionCreateOneSubscriptionFieldMissingResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Error returned when the subscription has missing field(s) (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"SubscriptionDeleteOneByUIDSubscriptionNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Error returned when the specified subscription does not exist (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"SubscriptionGetOneByUIDResponseBody":{"title":"SubscriptionGetOneByUIDResponseBody","type":"object","properties":{"email":{"type":"string","description":"Subscribed email","example":"user@email.com"},"name":{"type":"string","description":"Name of the subscriber","example":"John Doe"},"since":{"type":"string","description":"Datetime when the subscription was made","example":"2012-04-23T18:25:43.511Z"},"uid":{"type":"string","description":"Unique identifier of subcription","example":"abcd1234"}},"description":"A subscription with the same UID","example":{"email":"user@email.com","name":"John Doe","since":"2012-04-23T18:25:43.511Z","uid":"abcd1234"}},"SubscriptionGetOneByUIDSubscriptionNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Error returned when the specified subscription does not exist (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"SubscriptionResponse":{"title":"SubscriptionResponse","type":"object","properties":{"email":{"type":"string","description":"Subscribed email","example":"user@email.com"},"name":{"type":"string","description":"Name of the subscriber","example":"John Doe"},"since":{"type":"string","description":"Datetime when the subscription was made","example":"2012-04-23T18:25:43.511Z"},"uid":{"type":"string","description":"Unique identifier of subcription","example":"abcd1234"}},"description":"Representation of the subscription","example":{"email":"user@email.com","name":"John Doe","since":"2012-04-23T18:25:43.511Z","uid":"abcd1234"}}}}